Rename 'tasks' module to 'tasks-functions' and fix related imports to match
